[Enhancement] distribution property support null strict and null relax mode  (#24179)
Fixes #issue

## Core Points:
Utilize outer join calculation features, relax the requirements for
output properties of children, and reduce unnecessary data shuffling.

The purpose of shuffling is to make the data conform to certain
distribution characteristics. For example, after an `EXCHANGE SHUFFLE
[1, 4]`, the data will be strictly distributed based on columns 1 and 4.
To describe the data distribution more precisely, we can differentiate
the distribution behavior of null values. For instance, the data may
have been hash-distributed based on the values of a certain column, but
the null values in that column have been randomly allocated to various
buckets. For convenience, we denote data strictly distributed according
to a column as `colId(null strict)`, and data distributed according to
that column but with null values not conforming to the distribution
characteristics as `colId(null relax)`.

Let's look two examples.
#### Scenario 1:
`select * from t1 left join t2 on t1.v1 = t2.v1 left join t3 on t2.v1 =
t3.v1;`
The data distribution after the first join can be described as
`t1.v1(null strict)` or `t2.v1(null relax)`. Previous processing would
insert an exchange based on the join condition to strictly distribute
the results of the first join to `t2.v1(null strict)`.

However, according to the execution method of the left join, the rows
with null values in `t2.v1` have no distribution requirement for the
data in t3. Therefore, `t2.v1(null relax)` satisfies the requirement
without the need for additional shuffling.

#### Scenario 2:
`select * from t3 left join (select t2.v1 from t1 left join t2 on t1.v1
= t2.v1 group by t2.v1) t on t3.v1 = t.v1;`

The data distribution after `t1 left join t2` can be described as
`t1.v1(null strict)` or `t2.v1(null relax)`. To ensure the accuracy of
the aggregation result, the data needs to be reshuffled based on `t2.v1`
after local aggregation to ensure that it is globally aggregated
according to `t2.v1(null strict)`. Otherwise, when only use one phase
agg, the wrong result will contain multiple rows with `t2.v1 is null`.

However, when observing the outer left join and the join condition
`t3.v1 = t.v1`, it does not concern how many `t.v1 is null` rows in the
right table. In other words, the requirement for the output of the
aggregation can be relaxed at this moment, as long as the aggregation
result for non-null rows is accurate. Hence, only the results of the
first-stage aggregation rresult join with `t3` can ensure the final
correctness of the results.

## Main work
1. add null strict/relax info in distribution col
2. set null strict/relax info when derived required property, and output
property.
3. use a disjoint-set data structure `DistributionDisjointSet` to record
derived distribution cols from equalPredicate in on clause.
4. speical process when enforce.
- the `DistributionProperty` of an enforceOperator is always null
strict.
- the `distributionSpec` in `DistributionProperty` of an outputProperty
sholud always use the same object with the `distributionSpec` in
`PhysicalDistributionOperator`, or when prune the shuffle cols in
`PhysicalDistributionOperator`, the `requiredProperties` in an
optExpression won't change and it may affect the next process. So when
an group add the second enforcer, it should compare if there is a same
outputProperty enforcer with low cost. If yes, we use the exist low cost
enforcer and its outputProperty, if not, we use the new enforcer to
replace the old one.
```
public PhysicalPropertySet updateOutputPropertySet(GroupExpression expression, double cost,
    PhysicalPropertySet requiredPropertySet) {
    for (Map.Entry < PhysicalPropertySet, Pair < Double, GroupExpression >> entry: lowestCostExpressions.entrySet()) {
        if (entry.getKey().equals(requiredPropertySet)) {
            if (entry.getValue().first > cost) {
                lowestCostExpressions.put(requiredPropertySet, new Pair < > (cost, expression));
                return requiredPropertySet;
            } else {
                // already has an enforcer, we should use the origin one.
                return entry.getKey();
            }
        }
    }
    lowestCostExpressions.put(requiredPropertySet, new Pair < > (cost, expression));
    return requiredPropertySet;
}
```

---------

Signed-off-by: packy <wangchao@starrocks.com>
Signed-off-by: packy92 <wangchao@starrocks.com>
(cherry picked from commit 57266b36a0b3784afb11ab64d807fecc464701e4)